renamed handicap2021 to handicapWHS
